Boke out viewer code to viewer
