+++ server/src/main.rs
@@ -1,5 +1,6 @@
 mod auth;
 mod util;
+mod viewer;
 
 use auth::{login, regen_token, register_streamer};
 use axum::{
@@ -10,21 +11,16 @@ use axum::{
     routing::post,
     Json, Router,
 };
-use chrono::{DateTime, Utc};
-use rustis::{
-    client::Client,
-    commands::StringCommands,
-    commands::{GenericCommands, ListCommands},
-    Result,
-};
+use rustis::{client::Client, commands::GenericCommands, Result};
 use serde::{Deserialize, Serialize};
 use sqlx::postgres::{PgPool, PgPoolOptions};
 use std::env;
 use tower_http::cors::{Any, CorsLayer};
 use util::verify_hash;
+use viewer::{get_active_streamers, push, shift};
 
 #[derive(Deserialize, Debug)]
-struct Viewer {
+pub struct Viewer {
     user_id: String,
     input: String,
     stream: String,
@@ -65,7 +61,7 @@ pub struct AppRes<'a, T> {
 }
 
 #[derive(Deserialize, Serialize, Debug)]
-struct AppReq {
+pub struct AppReq {
     api_token: String,
     org_id: String,
 }
@@ -85,103 +81,6 @@ struct Streams {
 pub struct Login {
     org_id: String,
     password: String,
-}
-
-// TODO: change unwraps to expect
-
-async fn shift(
-    State(mut state): State<AppState>,
-    extract::Json(payload): extract::Json<AppReq>,
-) -> Json<AppRes<'static, Vec<String>>> {
-    let streamer =
-        sqlx::query_as::<_, Streamer>("SELECT * FROM streamers WHERE org_id = $1 LIMIT 1")
-            .bind(payload.org_id)
-            .fetch_one(&state.db_pool)
-            .await
-            .unwrap();
-
-    if verify_hash(&streamer.api_token.unwrap(), &payload.api_token) == false {
-        return Json(AppRes {
-            body: None,
-            error: Some("Error with validation"),
-        });
-    }
-
-    if streamer.active_stream.unwrap() == false {
-        return Json(AppRes {
-            body: None,
-            error: Some("Stream not active"),
-        });
-    }
-
-    let res: Vec<String> = state.redis_client.rpop(streamer.stream, 10).await.unwrap();
-    Json(AppRes {
-        body: Some(res),
-        error: None,
-    })
-}
-
-#[axum_macros::debug_handler]
-async fn push(
-    State(mut state): State<AppState>,
-    extract::Json(payload): extract::Json<Viewer>,
-) -> Json<AppRes<'static, &'static str>> {
-    let last_push: Option<String> = state
-        .redis_client
-        .get(payload.user_id.clone())
-        .await
-        .unwrap();
-
-    if last_push.is_some() {
-        if (Utc::now().time()
-            - last_push
-                .clone()
-                .unwrap()
-                .parse::<DateTime<Utc>>()
-                .unwrap()
-                .time())
-        .num_milliseconds()
-            < 1000
-        {
-            return Json(AppRes {
-                body: None,
-                error: Some(
-                    "Slow down you are sending too many inputs! \nYou can only send 1 a second.",
-                ),
-            });
-        }
-    }
-
-    let _ = state
-        .redis_client
-        .set(payload.user_id, Utc::now().to_string())
-        .await
-        .unwrap();
-
-    let streamer =
-        sqlx::query_as::<_, Streamer>("SELECT * FROM streamers WHERE stream = $1 LIMIT 1")
-            .bind(&payload.stream)
-            .fetch_one(&state.db_pool)
-            .await
-            .unwrap();
-
-    if streamer.active_stream.unwrap() == false {
-        return Json(AppRes {
-            body: None,
-            error: Some("Stream not active"),
-        });
-    }
-
-    let _ = state
-        .redis_client
-        .lpush(payload.stream, payload.input)
-        .await
-        .unwrap();
-
-    Json(AppRes {
-        body: Some("Success"),
-        error: None,
-    })
 }
 
 async fn toggle_stream(
@@ -220,17 +119,6 @@ async fn toggle_stream(
     Json(AppRes {
         body: Some(!streamer.active_stream.unwrap()),
         error: None,
-    })
-}
-
-async fn get_active_streamers(State(state): State<AppState>) -> extract::Json<Streams> {
-    Json(Streams {
-        names: sqlx::query_as::<_, Stream>(
-            "SELECT stream, org_id FROM streamers where active_stream = True",
-        )
-        .fetch_all(&state.db_pool)
-        .await
-        .unwrap(),
     })
 }
